Initial work on a framework for networking in facter.
This library is an attempt to replace the Facter::Util::Ip stuff inside
Facter with a new, improved, better attempt at handling network
information.  The previous refactorings were too bogged down in what
previously existed.

This is also a playground for me to experiment with OO more effectively
than in the past.

Warning:  Dragons lurk.